Phase 1: Move edit/editor based types into the edit entrypoint
