+++ packages/diffs/src/editor/matchBrackets.ts
@@ -1,6 +1,6 @@
-import type { Position, Range } from '../types';
 import type { TextDocument } from './textDocument';
 import type { EditorTokenizer } from './tokenizer';
+import type { Position, Range } from './types';
 
 const OPEN_BRACKETS = new Map([
   ['(', ')'],
