+++ packages/diffs/src/types.ts
@@ -13,7 +13,16 @@ import type {
   ThemeRegistrationResolved,
 } from 'shiki';
 
-import type { EditorEditCompleteEvent, EditState } from './editor/types';
+import type {
+  CapturedDiffSessionState,
+  EditorActiveLineOptions,
+  EditorCaret,
+  EditorChangeEvent,
+  EditorEditCompleteEvent,
+  EditorViewState,
+  EditState,
+  RetainedDiffSessionSnapshot,
+} from './editor/types';
 
 export type { CreatePatchOptionsNonabortable };
 
@@ -372,19 +381,6 @@ export interface FileDiffMetadata {
   cacheKey?: string;
 }
 
-/** FileDiff baseline and hunk state needed to resume an editing session. */
-export interface RetainedDiffSessionSnapshot {
-  oldFile: { name: string; lines: string[] } | null;
-  type: ChangeTypes;
-  hunks: Hunk[];
-}
-
-/** @internal Current diff state and whether the document changed while editing. */
-export interface CapturedDiffSessionState {
-  diffSession: RetainedDiffSessionSnapshot;
-  hasChanges: boolean;
-}
-
 export type MergeConflictMarkerRowType =
   | 'marker-start'
   | 'marker-base'
@@ -559,14 +555,6 @@ export type DiffLineAnnotation<T = undefined> = {
   side: AnnotationSide;
   lineNumber: number;
 } & OptionalMetadata<T>;
-
-/**
- * An edit-completion handler's decision: `'accept'` installs the completed
- * value the event carries, `'reject'` restores the external input. The event
- * is frozen, so re-key the accepted value in place (`event.file.cacheKey =
- * '…'`) before returning `'accept'`.
- */
-export type EditCompletionDecision = 'accept' | 'reject';
 
 export type CodeViewFileItem<T = undefined> = {
   id: string;
@@ -1010,13 +998,6 @@ export interface DiffsComponentOptions extends BaseCodeOptions {
   lineHoverHighlight?: 'disabled' | 'both' | 'number' | 'line';
 }
 
-export interface EditorActiveLineOptions {
-  lineNumberOnly?: boolean;
-  side?: SelectionSide;
-}
-
-export type EditorDocumentKind = 'file' | 'file-diff';
-
 export interface DiffsBaseComponent {
   readonly type: 'file' | 'file-diff' | 'unresolved-file';
   readonly top?: number;
@@ -1227,146 +1208,6 @@ export interface DiffsEditor<LAnnotation> {
   /** Replace the remote carets and selections. */
   setCarets(carets: EditorCaret<unknown>[]): void;
   cleanUp(reason?: 'discard' | 'recycle' | 'complete'): void;
-}
-
-/**
- * Position in a text document expressed as zero-based line and character offset.
- * The offsets are based on a UTF-16 string representation. So a string of the form
- * `a𐐀b` the character offset of the character `a` is 0, the character offset of `𐐀`
- * is 1 and the character offset of b is 3 since `𐐀` is represented using two code
- * units in UTF-16.
- *
- * Positions are line end character agnostic. So you can not specify a position that
- * denotes `\r|\n` or `\n|` where `|` represents the character offset.
- */
-export interface Position {
-  /**
-   * Line position in a document (zero-based).
-   *
-   * If a line number is greater than the number of lines in a document, it
-   * defaults back to the number of lines in the document.
-   * If a line number is negative, it defaults to 0.
-   *
-   * The above two properties are implementation specific.
-   */
-  readonly line: number;
-  /**
-   * Character offset on a line in a document (zero-based).
-   *
-   * The meaning of this offset is determined by the negotiated
-   * `PositionEncodingKind`.
-   *
-   * If the character value is greater than the line length it defaults back
-   * to the line length. This property is implementation specific.
-   */
-  readonly character: number;
-}
-
-/**
- * A range in a text document expressed as (zero-based) start and end positions.
- *
- * If you want to specify a range that contains a line including the line ending
- * character(s) then use an end position denoting the start of the next line.
- * For example:
- * ```ts
- * {
- *     start: { line: 5, character: 23 }
- *     end : { line 6, character : 0 }
- * }
- * ```
- */
-export interface Range {
-  /**
-   * The range's start position.
-   */
-  readonly start: Position;
-  /**
-   * The range's end position.
-   */
-  readonly end: Position;
-}
-
-/**
- * A text edit applicable to a text document.
- */
-export interface TextEdit {
-  /**
-   * The range of the text document to be manipulated. To insert
-   * text into a document create a range where start === end.
-   */
-  readonly range: Range;
-  /**
-   * The string to be inserted. For delete operations use an
-   * empty string.
-   */
-  readonly newText: string;
-}
-
-/** Different with `TextEdit`, the range has been resolved to offsets. */
-export interface ResolvedTextEdit {
-  /** The start offset of the text change. */
-  readonly start: number;
-  /** The end offset of the text change. */
-  readonly end: number;
-  /** The string to be inserted. For delete operations use an empty string. */
-  readonly text: string;
-}
-
-/** A normalized text change reported by the editor. */
-export interface EditorChange extends ResolvedTextEdit {
-  /** The replaced range in the document before the change. */
-  range: Range;
-}
-
-/** The document and normalized edits reported after an editor change. */
-export interface EditorChangeEvent<LAnnotation, TMode extends 'file' | 'diff'> {
-  changes: EditorChange[];
-  file: FileContents;
-  editor: DiffsEditor<LAnnotation>;
-  lineAnnotations?: TMode extends 'file'
-    ? LineAnnotation<LAnnotation>[]
-    : DiffLineAnnotation<LAnnotation>[];
-}
-
-/**
- * The direction of a selection.
- * -1: backward
- *  0: none
- *  1: forward
- */
-export type SelectionDirection = -1 | 0 | 1;
-
-export interface EditorSelection extends Range {
-  direction: SelectionDirection;
-}
-
-/** Visual metadata shared by a remote caret and its optional highlight. */
-export interface CaretMetadata {
-  /** CSS color used for the caret and its derived highlight tint. */
-  color: string;
-}
-
-/**
- * A non-editable, externally owned selection. This follows the browser's
- * anchor/focus model: matching positions render a caret, and differing
- * positions render a highlighted selection with its caret at `focus`.
- */
-export interface EditorCaret<T> {
-  anchor: Position;
-  focus: Position;
-  metadata: T & CaretMetadata;
-}
-
-export interface EditorViewportState {
-  /** Horizontal position owned by the current editable code scroller. */
-  scrollLeft: number;
-  /** Vertical position of the editor viewport. */
-  scrollTop?: number;
-}
-
-export interface EditorViewState {
-  selections?: EditorSelection[];
-  view?: EditorViewportState;
 }
 
 export interface DiffsTextDocument {
