+++ packages/diffs/src/utils/cloneFileDiffMetadata.ts
@@ -1,7 +1,6 @@
 import type {
   FileDiffMetadata,
   Hunk,
-  RetainedDiffSessionSnapshot,
 } from '../types';
 
 export function cloneHunks(hunks: Hunk[]): Hunk[] {
@@ -9,19 +8,6 @@ export function cloneHunks(hunks: Hunk[]): Hunk[] {
     ...hunk,
     hunkContent: hunk.hunkContent.map((content) => ({ ...content })),
   }));
-}
-
-export function cloneRetainedDiffSessionSnapshot(
-  snapshot: RetainedDiffSessionSnapshot
-): RetainedDiffSessionSnapshot {
-  return {
-    ...snapshot,
-    oldFile:
-      snapshot.oldFile != null
-        ? { ...snapshot.oldFile, lines: [...snapshot.oldFile.lines] }
-        : null,
-    hunks: cloneHunks(snapshot.hunks),
-  };
 }
 
 export function cloneFileDiffMetadata(
