+++ packages/diffs/test/editorSelection.test.ts
@@ -41,13 +41,13 @@ import {
 } from '../src/editor/selection';
 import { DirectionBackward } from '../src/editor/selection';
 import { TextDocument } from '../src/editor/textDocument';
-import { disposeHighlighter } from '../src/highlighter/shared_highlighter';
 import type {
   EditorSelection,
-  FileContents,
   ResolvedTextEdit,
   SelectionDirection,
-} from '../src/types';
+} from '../src/editor/types';
+import { disposeHighlighter } from '../src/highlighter/shared_highlighter';
+import type { FileContents } from '../src/types';
 import { installDom, wait } from './domHarness';
 
 afterAll(async () => {
