+++ packages/diffs/test/editorState.test.ts
@@ -3,13 +3,12 @@ import { describe, expect, test } from 'bun:test';
 import { Editor } from '../src/editor/editor';
 import { EditStateManager } from '../src/editor/EditStateManager';
 import { TextDocument } from '../src/editor/textDocument';
-import type { EditState } from '../src/editor/types';
+import type { EditorViewState, EditState } from '../src/editor/types';
 import type {
   DiffLineAnnotation,
   DiffsEditableComponent,
   DiffsEditor,
   DiffsHighlighter,
-  EditorViewState,
   FileContents,
   HighlightedToken,
   RenderRange,
